+++ packages/techdocs-cli/src/lib/PublisherConfig.test.ts
@@ -68,6 +68,7 @@ describe('getValidPublisherConfig', () => {
       const config = {
         publisherType: 'awsS3',
         storageName: 'someStorageName',
+        awsBucketRootPath: 'backstage-data/techdocs',
       } as unknown as Command;
 
       const actualConfig = PublisherConfig.getValidConfig(config);
@@ -75,6 +76,23 @@ describe('getValidPublisherConfig', () => {
       expect(
         actualConfig.getString('techdocs.publisher.awsS3.bucketName'),
       ).toBe('someStorageName');
+      expect(
+        actualConfig.getString('techdocs.publisher.awsS3.bucketRootPath'),
+      ).toBe('backstage-data/techdocs');
+    });
+
+    it('should return valid ConfigReader with SSE option', () => {
+      const config = {
+        publisherType: 'awsS3',
+        storageName: 'someStorageName',
+        awsS3sse: 'aws:kms',
+      } as unknown as Command;
+
+      const actualConfig = PublisherConfig.getValidConfig(config);
+      expect(actualConfig.getString('techdocs.publisher.type')).toBe('awsS3');
+      expect(actualConfig.getString('techdocs.publisher.awsS3.sse')).toBe(
+        'aws:kms',
+      );
     });
   });
 
@@ -127,6 +145,7 @@ describe('getValidPublisherConfig', () => {
       const config = {
         publisherType: 'googleGcs',
         storageName: 'someStorageName',
+        gcsBucketRootPath: 'backstage-data/techdocs',
       } as unknown as Command;
 
       const actualConfig = PublisherConfig.getValidConfig(config);
@@ -136,6 +155,9 @@ describe('getValidPublisherConfig', () => {
       expect(
         actualConfig.getString('techdocs.publisher.googleGcs.bucketName'),
       ).toBe('someStorageName');
+      expect(
+        actualConfig.getString('techdocs.publisher.googleGcs.bucketRootPath'),
+      ).toBe('backstage-data/techdocs');
     });
   });
 });