+++ packages/techdocs-cli/src/lib/PublisherConfig.ts
@@ -85,9 +85,11 @@ export class PublisherConfig {
       type: 'awsS3',
       awsS3: {
         bucketName: cmd.storageName,
+        ...(cmd.awsBucketRootPath && { bucketRootPath: cmd.awsBucketRootPath }),
         ...(cmd.awsRoleArn && { credentials: { roleArn: cmd.awsRoleArn } }),
         ...(cmd.awsEndpoint && { endpoint: cmd.awsEndpoint }),
         ...(cmd.awsS3ForcePathStyle && { s3ForcePathStyle: true }),
+        ...(cmd.awsS3sse && { sse: cmd.awsS3sse }),
       },
     };
   }
@@ -122,6 +124,7 @@ export class PublisherConfig {
       type: 'googleGcs',
       googleGcs: {
         bucketName: cmd.storageName,
+        ...(cmd.gcsBucketRootPath && { bucketRootPath: cmd.gcsBucketRootPath }),
       },
     };
   }