+++ plugins/home/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-home
 
+## 0.4.6
+
+### Patch Changes
+
+- a125278b81: Refactor out the deprecated path and icon from RouteRefs
+- Updated dependencies
+  - @backstage/core-components@0.7.4
+  - @backstage/core-plugin-api@0.2.0
+
 ## 0.4.5
 
 ### Patch Changes